Organaize code

+++ CrashEdit/Controls/NSFBox.cs
@@ -63,8 +63,8 @@ namespace CrashEdit
                 SelectedNode = NSFController.Node,
                 AllowDrop = true,
                 Font = new Font("Arial", 9F),
-                BackColor = Color.FromArgb(((int)(((byte)(38)))), ((int)(((byte)(38)))), ((int)(((byte)(38))))),
-                ForeColor = Color.FromArgb(((int)(((byte)(200)))), ((int)(((byte)(200)))), ((int)(((byte)(200)))))
+                BackColor = Color.FromArgb(38, 38, 38),
+                ForeColor = Color.FromArgb(200, 200, 200)
             };
             trvMain.Nodes.Add(NSFController.Node);
             trvMain.AfterSelect += new TreeViewEventHandler(trvMain_AfterSelect);
@@ -74,7 +74,7 @@ namespace CrashEdit
 
             pnSplit = new SplitContainer { Dock = DockStyle.Fill };
             pnSplit.Panel1.Controls.Add(trvMain);
-            pnSplit.BackColor = Color.FromArgb(((int)(((byte)(45)))), ((int)(((byte)(45)))), ((int)(((byte)(48)))));
+            pnSplit.BackColor = Color.FromArgb(45, 45, 48);
 
             Controls.Add(pnSplit);
         }
